Update package.json
When trying to compile the app on Node v17 I was getting errors error:0308010C:digital envelope routines::unsupported reading online showed this was related to openssl. The hacks and workarounds seem to disable security. But an easier option was to use Node v16 - so I have added a constraint to the node engine in the package.json"
[master e732cfa] When trying to compile the app on Node v17 I was getting errors error:0308010C:digital envelope routines::unsupported reading online showed this was related to openssl. The hacks and workarounds seem to disable security. But an easier option was to use Node v16 - so I have added a constraint to the node engine in the package.json

I also noticed that veraPDF version numbers didn't match, so made them match the current version that gets downloaded bu the download script